build script: change kernel.string for anykernel * now we know exactly which kernel we are installing * also a little refactoring

+++ build.sh
@@ -1,23 +1,25 @@
 #!/bin/bash
-arch="arm"
-KERNEL_DIR=$PWD
+
+export ARCH="arm"
 export KBUILD_BUILD_HOST="eOS-0.4.1-Loki"
 export KBUILD_BUILD_USER="arttttt"
-export ARCH=$arch
-clean_build=0
-config="tegra12_android_defconfig"
-original_output="$KERNEL_DIR/arch/arm/boot"
-output="$KERNEL_DIR/output"
+
 kernel_name="SmokeR24.2"
 threads=5
 toolchain="$HOME/Android/toolchain/linaro-4.9.4/bin/arm-linux-gnueabihf-"
 
+KERNEL_DIR=$PWD
+ORIGINAL_OUTPUT_DIR="$KERNEL_DIR/arch/$ARCH/boot"
+OUTPUT_DIR="$KERNEL_DIR/output"
+
 generate_version()
 {
-	if [ -f "$KERNEL_DIR/.git/HEAD" ]; then
+	if [[ -f "$KERNEL_DIR/.git/HEAD"  &&  -f "$KERNEL_DIR/anykernel/anykernel.sh" ]]; then
+		local updated_kernel_name
+		eval "$(awk -F"="  '/kernel.string/{print "anykernel_name="$2}' $KERNEL_DIR/anykernel/anykernel.sh)"
 		eval "$(awk -F"-"  '{print "current_branch="$2}' $KERNEL_DIR/.git/HEAD)"
-		if [ "$current_branch" != "stable" ]; then
-			if [ ! -f "$KERNEL_DIR/version" ]; then
+		if [[ "$current_branch" != "stable" ]]; then
+			if [[ ! -f "$KERNEL_DIR/version" ]]; then
 				echo "build_number=0" > $KERNEL_DIR/version
 			fi;
 
@@ -25,7 +27,11 @@ generate_version()
 			mv tmpfile $KERNEL_DIR/version
 			eval "$(awk -F"="  '{print "current_build="$2}' $KERNEL_DIR/version)"
 			export LOCALVERSION="-build$current_build"
+			updated_kernel_name=$kernel_name"-"$current_branch"-build"$current_build
+		else
+			updated_kernel_name=$kernel_name"-"$current_branch
 		fi;
+			sed -i s/$anykernel_name/$updated_kernel_name/ $KERNEL_DIR/anykernel/anykernel.sh
 	fi;
 }
 
@@ -34,14 +40,14 @@ make_zip()
 	printf "\nСоздание zip архива\n\n"
 
 	cd $KERNEL_DIR/anykernel
-	local zip_name="$kernel_name($(date +'%d.%m.%Y-%H:%M'))"
-	zip -r $zip_name.zip *
+	local zip_name="$kernel_name($(date +'%d.%m.%Y-%H:%M')).zip"
+	zip -r $zip_name *
 
-	if [ -f "$PWD/$zip_name.zip" ]; then
-		printf "\n$zip_name.zip создан, перемещение в $output"
-		mv "$PWD/$zip_name.zip" $output
+	if [[ -f "$PWD/$zip_name" ]]; then
+		printf "\n$zip_name создан, перемещение в $OUTPUT_DIR"
+		mv "$PWD/$zip_name" $OUTPUT_DIR
 
-		if [ -f "$output/$zip_name.zip" ]; then
+		if [[ -f "$OUTPUT_DIR/$zip_name" ]]; then
 			echo
 			printf "\nЗавершено\n"
 		fi
@@ -56,27 +62,27 @@ compile()
 	local start=$(date +%s)
 	clear
 
-	if [ "$clean_build" == 1 ]; then
+	if [[ "$clean_build" == 1 ]]; then
 		make clean
 		make mrproper
 	fi
 
 	generate_version
 	make $config
-	make -j$threads ARCH=$arch CROSS_COMPILE=$toolchain zImage
+	make -j$threads ARCH=$ARCH CROSS_COMPILE=$toolchain zImage
 
-	if [ -f "$original_output/zImage" ]; then
-		mv $original_output/zImage $PWD/anykernel/kernel/
+	if [[ -f "$ORIGINAL_OUTPUT_DIR/zImage" ]]; then
+		mv $ORIGINAL_OUTPUT_DIR/zImage $PWD/anykernel/kernel/
 	else
 		exit 0
 	fi
 
 	printf "\nКомпиляция дерева устройства\n\n"
 
-	make -j$threads ARCH=$arch CROSS_COMPILE=$toolchain tegra124-mocha.dtb
+	make -j$threads ARCH=$ARCH CROSS_COMPILE=$toolchain tegra124-mocha.dtb
 
-	if [ -f "$original_output/dts/tegra124-mocha.dtb" ]; then
-		mv $original_output/dts/tegra124-mocha.dtb $PWD/anykernel/kernel/dtb
+	if [[ -f "$ORIGINAL_OUTPUT_DIR/dts/tegra124-mocha.dtb" ]]; then
+		mv $ORIGINAL_OUTPUT_DIR/dts/tegra124-mocha.dtb $PWD/anykernel/kernel/dtb
 	else
 		exit 0
 	fi
@@ -92,11 +98,13 @@ compile()
 compile_dtb()
 {
 	clear
-	make $config
-	make -j$threads ARCH=$arch CROSS_COMPILE=$toolchain tegra124-mocha.dtb
 
-	if [ -f "$original_output/dts/tegra124-mocha.dtb" ]; then
-		mv $original_output/dts/tegra124-mocha.dtb $PWD/anykernel/kernel/dtb
+	generate_version
+	make $config
+	make -j$threads ARCH=$ARCH CROSS_COMPILE=$toolchain tegra124-mocha.dtb
+
+	if [[ -f "$ORIGINAL_OUTPUT_DIR/dts/tegra124-mocha.dtb" ]]; then
+		mv $ORIGINAL_OUTPUT_DIR/dts/tegra124-mocha.dtb $PWD/anykernel/kernel/dtb
 	else
 		exit 0
 	fi
